backend/requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# CTGAN for synthetic data generation
ctgan>=0.5.0

# API framework
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0

# Data processing
python-dateutil>=2.8.0
pytz>=2021.1

# Visualization and analysis
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Logging and utilities
python-dotenv>=0.19.0
tqdm>=4.62.0

# ====================
# Vector Database Libraries
# ====================
# ChromaDB (local, lightweight, default)
chromadb>=0.4.22

# Pinecone (cloud-based, scalable)
pinecone-client>=3.0.0

# Weaviate (self-hosted or cloud)
weaviate-client>=4.0.0

# FAISS (Facebook AI Similarity Search - local, fast)
faiss-cpu>=1.7.4

# ====================
# Embedding Libraries
# ====================
# Sentence Transformers (local embeddings)
sentence-transformers>=2.2.2

# OpenAI (API-based embeddings)
openai>=1.12.0

# HuggingFace Transformers (for BioBERT, PubMedBERT)
transformers>=4.30.0
torch>=2.0.0

# ====================
# LangChain Framework
# ====================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.3.0
langsmith>=0.1.125

# ====================
# Web Scraping Libraries
# ====================
requests>=2.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.8.0
feedparser>=6.0.0

# Selenium (for dynamic content)
selenium>=4.15.0
webdriver-manager>=4.0.0

# ====================
# Async and HTTP
# ====================
httpx>=0.25.0
tenacity>=8.2.0

# ====================
# Task Scheduling
# ====================
schedule>=1.2.0

# ====================
# Data Validation
# ====================
jsonschema>=4.0.0
